+++ src/Controller/StaffController.php
@@ -6,8 +6,8 @@
 
 namespace Incwadi\Core\Controller;
 
-use Incwadi\Core\Entity\Customer;
-use Incwadi\Core\Form\CustomerType;
+use Incwadi\Core\Entity\Staff;
+use Incwadi\Core\Form\StaffType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -15,9 +15,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/v1/customer", name="customer_")
+ * @Route("/v1/staff", name="staff_")
  */
-class CustomerController extends AbstractController
+class StaffController extends AbstractController
 {
     /**
      * @Route("/", methods={"GET"}, name="index")
@@ -27,7 +27,7 @@ class CustomerController extends AbstractController
     {
         return $this->json(
             [
-                'customers' => $this->getDoctrine()->getRepository(Customer::class)->findBy(
+                'staff' => $this->getDoctrine()->getRepository(Staff::class)->findBy(
                     [
                         'branch' => $this->getUser()->getBranch()
                     ]
@@ -40,9 +40,9 @@ class CustomerController extends AbstractController
      * @Route("/{id}", methods={"GET"}, name="show")
      * @Security("is_granted('ROLE_USER')")
      */
-    public function show(Customer $customer): JsonResponse
+    public function show(Staff $staff): JsonResponse
     {
-        return $this->json($customer);
+        return $this->json($staff);
     }
 
     /**
@@ -51,11 +51,11 @@ class CustomerController extends AbstractController
      */
     public function new(Request $request): JsonResponse
     {
-        $customer = new Customer();
-        $customer->setBranch(
+        $staff = new Staff();
+        $staff->setBranch(
             $this->getUser()->getBranch()
         );
-        $form = $this->createForm(CustomerType::class, $customer);
+        $form = $this->createForm(StaffType::class, $staff);
 
         $form->submit(
             json_decode(
@@ -65,14 +65,14 @@ class CustomerController extends AbstractController
         );
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
-            $em->persist($customer);
+            $em->persist($staff);
             $em->flush();
 
-            return $this->json($customer);
+            return $this->json($staff);
         }
 
         return $this->json([
-            'msg' => 'Please enter a valid customer!'
+            'msg' => 'Please enter a valid staff member!'
         ], 400);
     }
 
@@ -80,9 +80,9 @@ class CustomerController extends AbstractController
      * @Route("/{id}", methods={"PUT"}, name="edit")
      * @Security("is_granted('ROLE_ADMIN')")
      */
-    public function edit(Request $request, Customer $customer): JsonResponse
+    public function edit(Request $request, Staff $staff): JsonResponse
     {
-        $form = $this->createForm(CustomerType::class, $customer);
+        $form = $this->createForm(StaffType::class, $staff);
 
         $form->submit(
             json_decode(
@@ -94,11 +94,11 @@ class CustomerController extends AbstractController
             $em = $this->getDoctrine()->getManager();
             $em->flush();
 
-            return $this->json($customer);
+            return $this->json($staff);
         }
 
         return $this->json([
-            'msg' => 'Please enter a valid customer!'
+            'msg' => 'Please enter a valid staff member!'
         ]);
     }
 
@@ -106,14 +106,14 @@ class CustomerController extends AbstractController
      * @Route("/{id}", methods={"DELETE"}, name="delete")
      * @Security("is_granted('ROLE_ADMIN')")
      */
-    public function delete(Customer $customer): JsonResponse
+    public function delete(Staff $staff): JsonResponse
     {
         $em = $this->getDoctrine()->getManager();
-        $em->remove($customer);
+        $em->remove($staff);
         $em->flush();
 
         return $this->json([
-            'msg' => 'The customer was successfully deleted.'
+            'msg' => 'The staff member was successfully deleted.'
         ]);
     }
 }