+++ tests/Controller/AuthorTest.php
@@ -10,22 +10,17 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class AuthorTest extends WebTestCase
 {
-    protected $clientAdmin;
-
-    public function setUp(): void
-    {
-        $this->buildClient();
-    }
+    use \Incwadi\Core\Tests\ApiTestTrait;
 
     public function testScenario()
     {
         // find
-        $request = $this->request('/author/find', 'GET', ['term' => 'name']);
+        $request = $this->request('/v1/author/find', 'GET', ['term' => 'name']);
 
         $this->assertInternalType('array', $request->authors);
 
         // new
-        $request = $this->request('/author/new', 'POST', [], [
+        $request = $this->request('/v1/author/new', 'POST', [], [
             'firstname' => 'Firstname',
             'surname' => 'Surname'
         ]);
@@ -38,7 +33,7 @@ class AuthorTest extends WebTestCase
         $id = $request->id;
 
         // edit
-        $request = $this->request('/author/'.$id, 'PUT', [], [
+        $request = $this->request('/v1/author/'.$id, 'PUT', [], [
             'firstname' => 'Firstname1',
             'surname' => 'Surname1'
         ]);
@@ -49,7 +44,7 @@ class AuthorTest extends WebTestCase
         $this->assertEquals('Surname1', $request->surname);
 
         // show
-        $request = $this->request('/author/'.$id, 'GET');
+        $request = $this->request('/v1/author/'.$id, 'GET');
 
         $this->assertTrue(isset($request->id));
         $this->assertInternalType('integer', $request->id);
@@ -57,54 +52,8 @@ class AuthorTest extends WebTestCase
         $this->assertEquals('Surname1', $request->surname);
 
         // delete
-        $request = $this->request('/author/'.$id, 'DELETE');
+        $request = $this->request('/v1/author/'.$id, 'DELETE');
 
         $this->assertEquals('The author was successfully deleted.', $request->msg);
-    }
-
-    protected function request(string $url, ?string $method = 'GET', ?array $params = [], ?array $content = [])
-    {
-        $client = $this->clientAdmin;
-
-        $crawler = $client->request(
-            $method,
-            '/v1'.$url,
-            $params,
-            [],
-            [],
-            json_encode($content)
-        );
-
-        $this->assertTrue($client->getResponse()->isSuccessful(), 'Unexpected HTTP status code for '.$method.' '.$url.'!');
-
-        return json_decode($client->getResponse()->getContent());
-    }
-
-    protected function buildClient()
-    {
-        $this->clientAdmin = static::createClient();
-        $this->clientAdmin->request(
-            'POST',
-            '/api/login_check',
-            [],
-            [],
-            [
-                'CONTENT_TYPE' => 'application/json'
-            ],
-            json_encode(
-                [
-                    'username' => 'admin',
-                    'password' => 'password'
-                ]
-            )
-        );
-        $data = json_decode(
-            $this->clientAdmin->getResponse()->getContent(),
-            true
-        );
-        $this->clientAdmin->setServerParameter(
-            'HTTP_Authorization',
-            sprintf('Bearer %s', $data['token'])
-        );
     }
 }