+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.1.0
+
+- feature: add info whether the user has role user and admin
+- feature: rename customer to staff
+- refactor: rename UserFixtures to AppFixtures
+- build: add linters
+- feature: add trait for tests
+- feature: JWT Refresh Token
+
 ## 1.0.0
 
 - First Release